Improve visualization of PCA results
The plot showing results of PCA are now interactive - when hovering over
the points with a mouse, the index and molecular formula are displayed
in a yellow box, while the 1D scan for the molecule is displayed in a
figure below the main plot. This allows us to better understand what PCA
is doing.