Improve 1D scan visuals and set up clustering
Refactored the visualization so it is possible to plot clusters with
different colors. Furthermore, the 1D scan now shows up next to the
scatter plot, and it is larger and has more labels.

Also ran a basic clustering with AgglomerativeClustering from sklearn;
this demonstrates how clustering can be visualized in the future.