Display clusters
There are now visualizations of the 1D scans of molecules in each
cluster. This should make it easy to tell if the clusters seem to be
doing well. With the setting of 15 clusters in this commit, it seems
most clusters have similar 1D scans, except for cluster 2, which seems
to have some odd ones out.