Mark bad entries
Entries can now be easily marked as bad via a global `IS_GOOD_ENTRY`
array. This makes it easy to find molecules that are okay to use.

Also refined the corner detection - I manually checked each plot and
compared the results to the automatic detection. It's far from perfect,
but perhaps tolerable.

Visualization and clustering still remain to be updated.