Refactor Notebook
Refactored large portions of the notebook. The data is now represented
as one global list that is passed around between various functions,
instead of a bunch of objects that are modified by each function. This
makes things much easier to work with.

Also added the ability to save grid optimization data locally and load
it back into the original objects.

While doing all this, also accounted for:
- The new set of grid optimization data from the most recent run
- Bad data in the runs, courtesy of Trevor Gokey

Future changes will tidy up the whole notebook, and make the clustering
portion work properly.